+++ frontend/README.md
@@ -1,8 +1,5 @@
-# React + Vite
+Команды для консоли:
 
-This template provides a minimal setup to get React working in Vite with HMR and some ESLint rules.
-
-Currently, two official plugins are available:
-
-- [@vitejs/plugin-react](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react/README.md) uses [Babel](https://babeljs.io/) for Fast Refresh
-- [@vitejs/plugin-react-swc](https://github.com/vitejs/vite-plugin-react-swc) uses [SWC](https://swc.rs/) for Fast Refresh
+npm install -> загрузит папку node_modules
+npm run dev -> запустит локальный сервер для разработки
+npm run build -> сбилдит приложение
