+++ frontend/index.html
@@ -2,9 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="./user-icon.svg" />
+    <link rel="icon" type="image/svg+xml" href="/icons/user-icon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Registration Form (example)</title>
+    <title>Registration Form</title>
   </head>
   <body>
     <div id="root"></div>
