+++ frontend/src/components/forms/PasswordReset.jsx
@@ -1,0 +1,5 @@
+function PasswordReset() {
+  return <div>Password reset form will be there</div>;
+}
+
+export default PasswordReset;
